Separated action functions into the systems which implement them.
