+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL1/include.h
@@ -11,4 +11,4 @@
 #ifdef TARGET_OS_MAC
 #include "../General/GLsurface.h"
 #endif
-
+#include "../General/actions.h"
