+++ ENIGMAsystem/SHELL/Universal_System/Extensions/Paths/actions.h
@@ -1,0 +1,51 @@
+/********************************************************************************\
+ **                                                                              **
+ **  Copyright (C) 2010 Alasdair Morrison <tgmg@g-java.com>                      **
+ **  Copyright (C) 2011 Polygone                                                 **
+ **                                                                              **
+ **  This file is a part of the ENIGMA Development Environment.                  **
+ **                                                                              **
+ **                                                                              **
+ **  ENIGMA is free software: you can redistribute it and/or modify it under the **
+ **  terms of the GNU General Public License as published by the Free Software   **
+ **  Foundation, version 3 of the license or any later version.                  **
+ **                                                                              **
+ **  This application and its source code is distributed AS-IS, WITHOUT ANY      **
+ **  WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS   **
+ **  FOR A PARTICULAR PURPOSE. See the GNU General Public License for more       **
+ **  details.                                                                    **
+ **                                                                              **
+ **  You should have recieved a copy of the GNU General Public License along     **
+ **  with this code. If not, see <http://www.gnu.org/licenses/>                  **
+ **                                                                              **
+ **  ENIGMA is an environment designed to create games and other programs with a **
+ **  high-level, fully compilable language. Developers of ENIGMA or anything     **
+ **  associated with ENIGMA are in no way responsible for its users or           **
+ **  applications created by its users, or damages caused by the environment     **
+ **  or programs made in the environment.                                        **
+ **                                                                              **
+ \********************************************************************************/
+
+extern bool argument_relative;
+
+namespace enigma_user
+{
+
+inline void action_path(unsigned pathid,double speed,unsigned endaction,bool absolute)
+{
+    path_start(pathid, speed, endaction, absolute);
+}
+inline void action_path_end()
+{
+    path_end();
+}
+inline void action_path_position(double position, bool relative)
+{
+    path_set_position(position, relative);
+}
+inline void action_path_speed(double speed, bool relative)
+{
+    path_set_speed(speed, relative);
+}
+
+}
